mysite/requirements.txt
django==2.1.3
djangorestframework>=3.9.1, <3.10.0
#psycopg2>=2.7.5,<2.8.0
flake8>=3.6.0,<3.7.0
django-crispy-forms
mysqlclient
pymysql
django-admin-tools
gunicorn
whitenoise
adal>=0.4
requests>=2.20.0
passlib>=1.7.1
pycryptodomex>=3.9.6
pycryptodome>=3.9.6
django-webpack-loader>=0.7.0
djangorestframework-jwt>=1.11.0
